CHROMIUM: iwl7000: mvm: clear radiotap vendor data padding
We add some padding to the radiotap vendor data so that the
whole radiotap length is a multiple of 4, so we have proper
alignment of the IP header later. I evidently forgot to clear
this padding, leaking two bytes of memory contents. Clear it.

Change-Id: Iddfa17721104e9ff6bf02faa558c84350ee4a9c6
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 20080e28fd16620f68cc0112e51762c0673e4c6b