UPSTREAM: ASoC: core: Allow soc_find_component lookups to match parent of_node
For devices implemented as a MFD it is common to only have a single node
in devicetree representing the whole device. As such when looking up
components in soc_find_components we should match against both the devices
of_node and the devices parent's of_node, as is already done in the rest
of the ASoC core.

This causes regressions for some DAI links at the moment as
soc_find_component was recently added as a check in soc_init_dai_link.

Fixes: 8780cf1142a5 ("ASoC: soc-core: defer card probe until all component is added to list")
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit d0b95e6cd298a785c126e75a085af6dd7b7b1f60)

BUG=b:130538528
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: I11e0d3cc5e517f5a52fa4ead06419ac9268e3664
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>