CHROMIUM: ath10k: wait for vdev delete response from firmware
When we add an interface immediately after removing
the interface the vdev deletion in firmware might not
have been completed. We need to synchronize the vdev creation
with the firmware.

Wait for vdev delete response from firmware when we
remove an interface.

Tested HW: WCN3990
Tested FW: WLAN.HL.2.0-01188-QCAHLSWMTPLZ-1

Change-Id: I712f8dbc624e54fcdc776051fa73926beceea195
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1436721
Reviewed-by: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>