CHROMIUM: iwl7000: mvm: Allow retries for probe responses
When hostapd sends a probe response, it does not request an
acknowledgment on the probe response. This setting is propagated
to the driver and FW from mac80211, and as a result probe responses
are transmitted only once without retries. In congested
environments, this reduces the likelihood for finding the AP.

To overcome this, enable acknowledgment and retries for probe
responses, even in case no acknowledgment is requested. In such
a case the retry limit is set to 3.

Change-Id: I0c57839418e2a2ad23bc2dc3ba896a68f07bfccb
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: adba8c61beadfe068b974ba5aa3572f006ea22ac