CHROMIUM: iwl7000: fix send hcmd timeout recovery flow
In an earlier commit there was an attempt to fix the same issue.
It was made under the assumption that the driver always receives
an interrupt from the FW after generating a NMI. This assumption
is wrong, and in cases the driver does not receive an interrupt,
it does not start recovery flow.

Solve this by waiting for the FW interrupt.
if we get a timeout, the driver will start the recovery flow

Change-Id: I222a916e683de534e2753196d497658c3bdf5a99
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
iwl7000-tree: 82762564180c0821c2353a8df7f6d6c9b7441a01