CHROMIUM: Bluetooth: Fix to Set the BD address
This fixes the BT ON issue if no entry in DT for BD address,
means not writing BD address to VPD. So if no entry for BD address
in DT table, it will set the default BD address and setup
the device in proper state.

BUG=b:123314957
TEST=1) Successful compilation
     2) Verified by multiple reboots of device

Note: These Changes are temporary fixes and shall be reverted on
root causing.

Change-Id: I13e07e206dfeaa450fe3ab6bead51c88c6fa74fd
Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1451677
Reviewed-by: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: NARAYANRADDI MASTI <nmasti@google.com>