CHROMIUM: Bluetooth: hci_core: Fix to Set the BD address
This fixes the BT ON issue if no entry in DT for BD address,
means not writing BD address to VPD.
it will set the default BD address and setup
the device in proper state.

BUG=b:123314957
TEST=1) Successful compilation
     2) Verified by multiple reboots of device

Note: These Changes are temporary fixes and shall be reverted on
root causing.

Change-Id: I0615f3109194091ce80c45aafef158c4ed588761
Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1451678
Reviewed-by: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: NARAYANRADDI MASTI <nmasti@google.com>