UPSTREAM: ASoC: dapm: fix use-after-free issue with dailink sname
Commit 7620fe9161ce ("ASoC: topology: fix memory leak in
soc_tplg_dapm_widget_create") fixed a memory leak issue, but
additional tests and KASAN reports show a use-after-free in soc-dapm.

The widgets are created with a kmemdup operating on a template. The
"name" string is also duplicated, but the "sname" string is not. As a
result, when the template is freed after widget creation, its sname
string is still used.

Fix by explicitly duplicating the "sname" string, and freeing it when
required.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 199ed3e81c49a621ce6fcb630ab9f30d92db6718)

BUG=b:130538528
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: I666a5587708f5b15036e2e512141537deac62f2c
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>