CHROMIUM: Bluetooth: Remove BT chip power off for every BT OFF
This fixes the BT ON issue like in boot up it is trying to
do device up and down and again it do device up. few times FW
download fialing, so we are blocking the device down and
second device up in bootup. So by the time  device bootup
BT device up and running.

BUG=b:123314957
TEST=1) Successful compilation
     2) Verified by multiple reboots of device

Note: These Changes are temporary fixes and shall be reverted on
root causing.

Change-Id: I720cb53f94234888437879b8f4c31f0366722283
Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1451679
Reviewed-by: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: NARAYANRADDI MASTI <nmasti@google.com>