CHROMIUM: Fix Bluetooth controller event mask config during system suspend
As part of CL:1383763, a few Bluetooth controller events were disabled on
suspension, to prevent the system from waking up spuriously.
This change is applying the same event mask change as in CL:1383763, but
instead of overriding the entire set of events mask it selectively masks
out or restores the required events during suspend and resume. In addition,
the new Set Event Mask operation introduced in this change is using the
BlueZ/NewBlue new suspend and resume infrastucture.

As part of suspend/resume operations, BlueZ sends a new MGMT command to
the kernel to mask out/restore BT specified events. The kernel adds new
functions to handle the Set Event Mask request. The request contains a
value and mask parameters to enable overriding only selected events.

This change is for kerenl v4.4 only.

For more details on the
Bluetooth/NewBlue suspend infra refer to CL:1454328

BUG=chromium:930973
TEST=suspend and resume of system and verify event mask is set properly by
the kernel. Verify event mask HCI and MGMT messages in BTMON.
Signed-off-by: Yoni Shavit <yshavit@chromium.org>

Change-Id: I8168f5451b61ae8d13cfd66d4924623990c506be
Reviewed-on: https://chromium-review.googlesource.com/1464385
Commit-Ready: Yoni Shavit <yshavit@chromium.org>
Tested-by: Yoni Shavit <yshavit@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>