CHROMIUM: iwl7000: mvm: use do_div in ftm to avoid problems with 32-bit platforms
There is an issue with some 64-bit divisions on 32-bit platforms that
cause the following warning:

ERROR: "__divdi3" [drivers/net/wireless/intel/iwlwifi/mvm/iwlmvm.ko] undefined!

This is because gcc uses a helper function that is implemented in one
of its libraries that is not linked in the kernel.

To prevent this issue, we should use the do_div() helper function when
making divisions on 64-bit values.  Do this in
iwl_mvm_debug_range_resp() to avoid the problem.

Change-Id: Icf85b9f7f8a27c4133891b896b6aa81670fe1723
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 191dd32d5486b26824b75d0cc8e8b299265fcc6a