Revert "BACKPORT: FROMLIST: Use LDJ_DEVICE macro for existing Logitech mice"
This reverts commit b6e4d15c2cfb3a6450f82d3736f86329ab468df8.

The initial high-resolution scrolling patches have been reverted in
mainline kernel, as it turns out the API in them wasn't compatible with
some Microsoft mice, so we decided to start again. Instead of
backporting the revert commits, I'm simply reverting the original
backport commits, as this reduces the number of nasty merge conflicts.

The equivalent mainline revert commit for this one is at
https://patchwork.kernel.org/patch/10692687/.

BUG=chromium:888172
TEST=Checked that a Logitech Anywhere MX no longer outputs _HI_RES events.

Change-Id: Ie1c49fc2bb877c8051b031ee9c87053c653085f4
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1484369
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>