Revert "BACKPORT: FROMLIST: Enable high-resolution scrolling on Logitech mice"
This reverts commit 3802c96bc03683556818be698d5997803ce374fc.

The initial high-resolution scrolling patches have been reverted in
mainline kernel, as it turns out the API in them wasn't compatible with
some Microsoft mice, so we decided to start again. Instead of
backporting the revert commits, I'm simply reverting the original
backport commits, as this reduces the number of nasty merge conflicts.

The equivalent mainline revert commit for this one is at
https://patchwork.kernel.org/patch/10692679/.

BUG=chromium:888172
TEST=Checked that a Logitech Anywhere MX no longer outputs _HI_RES events.

Change-Id: I5c8cbe93dc1613b8a1eeea002095eb327170a29a
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1484370
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>