Revert "BACKPORT: FROMLIST: Add function to enable HID++ 1.0 "scrolling acceleration""
This reverts commit 1cd37b5a86de0ebe4178e228ccdb3bff310bb9da.

The initial high-resolution scrolling patches have been reverted in
mainline kernel, as it turns out the API in them wasn't compatible with
some Microsoft mice, so we decided to start again. Instead of
backporting the revert commits, I'm simply reverting the original
backport commits, as this reduces the number of nasty merge conflicts.

The equivalent mainline revert commit for this one is at
https://patchwork.kernel.org/patch/10692685/.

BUG=chromium:888172
TEST=Checked that a Logitech Anywhere MX no longer outputs _HI_RES events.

Change-Id: I349a840d3ea330ad498dad8944bff41cb0a6292f
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1484891
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>