Revert "FROMLIST: Create a utility class for counting scroll events"
This reverts commit 50609688e5d62a388c6aa6e5168fe5635eac9a73.

The initial high-resolution scrolling patches have been reverted in
mainline kernel, as it turns out the API in them wasn't compatible with
some Microsoft mice, so we decided to start again. Instead of
backporting the revert commits, I'm simply reverting the original
backport commits, as this reduces the number of nasty merge conflicts.

The equivalent mainline revert commit for this one is at
https://patchwork.kernel.org/patch/10692683/.

BUG=chromium:888172
TEST=Checked that a Logitech Anywhere MX no longer outputs _HI_RES events.

Change-Id: Ia28d58693946718c0c5b91590a48e0125b1d5942
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1484892
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>