Revert "FROMLIST: Add the `REL_WHEEL_HI_RES` event code"
This reverts commit 23eddb2b670ab487598cb51ff157deb4ad707ca1.

The initial high-resolution scrolling patches have been reverted in
mainline kernel, as it turns out the API in them wasn't compatible with
some Microsoft mice, so we decided to start again. Instead of
backporting the revert commits, I'm simply reverting the original
backport commits, as this reduces the number of nasty merge conflicts.

The equivalent mainline revert commit for this one is at
https://patchwork.kernel.org/patch/10692681/.

BUG=chromium:888172
TEST=Checked that a Logitech Anywhere MX no longer outputs _HI_RES events.

Change-Id: I2b9267a72ccfc13542b8f3f8966651ab4498d396
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1484893
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>