UPSTREAM: ASoC: Intel: Add Geminilake Dialog Maxim machine driver
This patch enables support for GeminiLake with the DA7219 codec and
MAX98357A amplifier. To avoid duplicating code, the existing machine
driver for ApolloLake is reused with only changes in hardware
connectivity (SSP2 for DA7219 and SSP1 for MAX98357A).

The dailinks are directly modified in this patch. Using a helper would
be nicer, but it'll be done in a follow-up step with validation done
across multiple machine drivers.

Acked-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Yong Zhi <yong.zhi@intel.com>
Signed-off-by: Naveen Manohar <naveen.m@intel.com>
Signed-off-by: Harsha Priya <harshapriya.n@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit c011245a197017f8e9e9d140b658bdb2b702a0c5)

BUG=b:130538528
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: Iff253c632ff2d56067d5a5a2a7c42eaf99e1c276
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>