BACKPORT: arm64: errata: Pass --fix-cortex-a53-843419 to ld if workaround enabled
Cortex-A53 erratum 843419 is worked around by the linker, although it is
a configure-time option to GCC as to whether ld is actually asked to
apply the workaround or not.

This patch ensures that we pass --fix-cortex-a53-843419 to the linker
when both CONFIG_ARM64_ERRATUM_843419=y and the linker supports the
option.

Acked-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit 6ffe9923f2350c19b95a2c9ebf1b4f5f275986f2)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
[djkurtz: Skipping ARM64_MODULE_CMODEL_LARGE changes because
 chromeos-3.18 does not have commit fd045f6cd98 ("arm64: add support for
 module PLTs").]

# Conflicts:
#	arch/arm64/Kconfig

BUG=chromium:929415
TEST=Build and boot test hana; tryjob

Change-Id: I8ef0975a12d038b3cc1dcd9fcc96e13804d47998
Reviewed-on: https://chromium-review.googlesource.com/1471157
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>