UPSTREAM: media: ipu3: shut up warnings produced with W=1
There are lots of warnings produced by this driver. It is not
as much as atomisp, but it is still a lot.

So, use the same solution to hide most of them.
Those need to be fixed before promoting it out of staging,
so add it at the TODO list.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>

(cherry picked from commit 0bdfc56c13c0
git://linuxtv.org/media_tree.git master)

BUG=b:139647343
TEST=Sanity checked basic camera functions.

Change-Id: I253c1dcd241420634e6362ff7d1c1d6ff313b056
Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1624628
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>