CHROMIUM: add guado_moblab to 3.14 pre-cq
This is in the PRE_CQ_DEFAULT_CONFIGS

https://chromium.googlesource.com/chromiumos/chromite/+/d7810fd2d7b31fe57f8a155653d94976223d68d0/lib/constants.py#633

but not here. The broadwell kernel is mostly similar as to samus-*,
but moblab does make a few USE flag changes that are worth covering
here.

This may help enable us to reduce the number of pre-cqs
(PRE_CQ_DEFAULT_CONFIGS) run for non-kernel CLs.

BUG=none
TEST=none

Change-Id: I49005149644c48ae8c0d9731e34af9adff32a0a5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1478046
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>