FROMLIST: regulator: fan53555: Add support for FAN53526
The FAN53526 differs from the FAN53555 only in that the mode bit in
VSEL0/VSEL1 is moved to the CONTROL register, the voltage selector mask
is extended by 1 bit and the step is different.

So extend the existing fan53555 driver to support FAN53526 as well.

BUG=b:124524751
TEST=Stepped through each frequency and observed voltage change

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(am from https://patchwork.kernel.org/patch/10822977/)
Change-Id: Idd36db860cb69053c60b782c2b9058b54dd7a726
Reviewed-on: https://chromium-review.googlesource.com/c/1480434
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>