CHROMIUM: iwl7000: pcie: don't crash on invalid RX interrupt
If for some reason the device gives us an RX interrupt before we're
ready for it, perhaps during device power-on with misconfigured IRQ
causes mapping or so, we can crash trying to access the queues.

Prevent that by checking that we actually have RXQs and that they
were properly allocated.

Change-Id: I9788268d93dfd7b4bfc7fd5e861e2312d4617cd8
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3ea164b2220008f0bbe5d327a3c39e2e177dc1fa