Revert "CHROMIUM: config: Enable MEMCG on IPQ40XX"
This reverts commit efad517219989098653fc1aca21c7035fb621434.

Reason for revert:
We found issues with running cgroup rules daemon (cgrulesengd) inside
minijail very close to the branch cut date and hence decided to hold
back the feature from releasing. Due to this, we may not do any
targeted testing on cgroups in M74. Just enabling cgroups memcg
creates a control group and places all the process in it,
which is a different behavior than not having it enabled.
Though leaving this feature enabled might provide us evidence about
the stability of the feature independent of its use, not having a
targeted testing might let bugs with cgroups (if any) go undiscovered.
Since no other feature needs this, we think it might be best to disable
it until we actually use it. since this feature is not shipping in M74,
we are undoing the change

Original change's description:
> CHROMIUM: config: Enable MEMCG on IPQ40XX
>
> Enabled MEMCG for ipq40XX platform with following command:
> ./chromeos/scripts/kernelconfig editconfig
>
> BUG=b:124001088
> TEST=Created a new image,flashed and checked the memory
> controller folders are present on the router or not.
>
> Autotest Job links:
> https://jetstream-moblab1.corp.google.com/afe/#tab_id=view_job&object_id=18923
> https://jetstream-moblab1.corp.google.com/afe/#tab_id=view_job&object_id=18928
>
> Change-Id: Ia75041bcec731bdaf3f85071a1f65d2062a7d70e
> Signed-off-by: Sivarama Krishna <eluris@chromium.org>
> Reviewed-on: https://chromium-review.googlesource.com/1464092
> Commit-Ready: Sivarama Krishna Eluri <eluris@google.com>
> Tested-by: Sivarama Krishna Eluri <eluris@google.com>
> Reviewed-by: Suresh Rajashekara <sureshraj@chromium.org>
> Reviewed-by: Grant Grundler <grundler@chromium.org>

BUG=b:124001088
Change-Id: I2cd363bd811afe84a94246f309b4674476640953
Signed-off-by: Sivarama Krishna <eluris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1510579
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sivarama Krishna Eluri <eluris@google.com>
Reviewed-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>