HACK: CHROMIUM: rk3399-vpu: Compute plane address with data_offset (jpege)
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.
This is HACK to abuse data_offset for the purpose.
This patch should be reverted after the appropriate way is defined.

BUG=chromium:901264, b/127895658
TEST=None

Change-Id: If8f50e4caf7ac9f050140332ed4b97a0ce4cd43c
Signed-off-by: Wei Lee <wtlee@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1510919
Commit-Ready: Wei Lee <wtlee@chromium.org>
Tested-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>