CHROMIUM: media: rockchip-vpu: Clear read cache before decoding
The VDEC IP block includes a read cache, which is not coherent with
writes coming out of the same block or from any other bus masters,
including the CPU. Moreover, the cache is IOVA addressed, so when
another buffer ends up mapped at the same IOVA, the cache could have
stale entries for the old buffer.

To fix the issue, the cache needs to be cleared. Ideally we would do
that only if the buffer is dirty or a remap happened. However, clearing
every frame did not lead to any performance effect above the measurement
error, so let's just keep it simple.

BUG=chromium:908314
BUG=b:123677023
TEST=video_VideoDecodeAccelerator.vp9.import on Kevin

Change-Id: I2e2fd850c574c65149df68dfe4c4db2c84a3adbc
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
[tfiga: Added proper commit message.]
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1517573
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>