CHROMIUM: iwl7000: dbg_ini: check debug TLV type explicitly
In ini debug TLVs bit 24 is set. The driver relies on it in the memory
allocation for the debug configuration. This implementation is
problematic in case of a new debug TLV that is not supported yet is added
and uses bit 24. In such a scenario the driver allocate space without
using it which causes errors in the apply point enabling flow.

Solve it by explicitly checking if a given TLV is part of the list of
the supported ini debug TLVs.

Change-Id: I860449a65feb292141170303dc19a441302b4242
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
iwl7000-tree: 81ac4274b05d6abfae0bdf7dc7392f729262b31f