BACKPORT: UPSTREAM: mtd: add sysfs nodes for querying the flash name and id
Currently, we don't have sysfs nodes for querying the underlying flash
name and flash id. This information is important especially when we
want to know the flash detail of the defective system. In order to
support the query, we add two pointers (*flashname, *id) into the
mtd_info structure and create two sysfs nodes (flashname, id). This
patch is modified based on the SPI-NOR flash system as we only have
that system now. But the idea should be applied to the other flash
driver like NAND flash.

BUG=b:132567445
TEST=Verify flashname and id are available under sysfs mtd

Change-Id: I911a7e2e337c86fccd438b13f995190a9027e8e0
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Signed-off-by: Yuji Sasaki <sasakiy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1616405
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>