CHROMIUM: Bluetooth: Enable BLuetooth LE splitter for x86-64
With this change, BT splitter can be used and newblue can be run. By
default the functionality of the splitter is disabled (CL:1352383) and
users can enable it by using the crosh command (CL:1121542).

For now x86-64 is prioritized and this will be enabled on all other
architectures as well later on.

BUG=chromium:859239,chromium:909045
TEST=With this built on nami, enable and disable newblue and check
that splitter is indeed enabled and disabled. Check that by default
(splitter disabled) Bluetooth works normally.

Change-Id: If57c803df3937a815a6418b3592733dfd96969e2
Signed-off-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1547991