CHROMIUM: media: rockchip-vpu: fix stateless API support
The stateless codec API specifies that the frames resolution must be set
on the OUTPUT queue, and then reflected to the CAPTURE queue as the
coded size. The driver however currently expects the resolution to be
set on the CAPTURE queue by the client. This is a violation of the
stateless API, so fix it.

BUG=chromium:948534
TEST=VDA unittest on Kevin passes.

Change-Id: I6939323e4828cd169b6bf53d5ce8e09b73ab62d0
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1547641
Reviewed-by: Tomasz Figa <tfiga@chromium.org>