UPSTREAM: bpf: restrict map value pointer arithmetic for unprivileged
commit 0d6303db7970e6f56ae700fa07e11eb510cda125 upstream.

Restrict map value pointer arithmetic for unprivileged users in that
arithmetic itself must not go out of bounds as opposed to the actual
access later on. Therefore after each adjust_ptr_min_max_vals() with a
map value pointer as a destination it will simulate a check_map_access()
of 1 byte on the destination and once that fails the program is rejected
for unprivileged program loads. We use this later on for masking any
pointer arithmetic with the remainder of the map value space. The
likelihood of breaking any existing real-world unprivileged eBPF
program is very small for this corner case.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit afb711a6e8d003eead79432c026f6654ae303fcb)
Signed-off-by: Xuewei Zhang <xueweiz@google.com>

BUG=b:131266440
TEST=None

Change-Id: I2a4a275c8e11474a9aaf6032afe7cde5375f285f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1585206
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Xuewei Zhang <xueweiz@google.com>
Tested-by: Xuewei Zhang <xueweiz@google.com>