CHROMIUM: media: rk3288-vpu: fix stateless API support
The stateless codec API specifies that the frames resolution must be set
on the OUTPUT queue, and then reflected to the CAPTURE queue as the
coded size. The driver however currently expects the resolution to be
set on the CAPTURE queue by the client. This is a violation of the
stateless API, so fix it.

BUG=chromium:948534
TEST=VDA unittest on Minnie passes with both unfixed and fixed Chromium.

Change-Id: Id67acdb35b726caefbb47158e6acd9dfdede81e2
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1553227
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>