FROMLIST: mfd: cros_ec: check for NULL transfer function
As new transfer mechanisms are added to the EC codebase, they may
not support v2 of the EC protocol.

If the v3 initial handshake transfer fails, the kernel will try
and call cmd_xfer as a fallback. If v2 is not supported, cmd_xfer
will be NULL, and the code will end up causing a kernel panic.

Add a check for NULL before calling the transfer function, along
with a helpful comment explaining how one might end up in this
situation.

Signed-off-by: Enrico Granata <egranata@chromium.org>

(from https://lkml.org/lkml/2019/4/3/1192)

BRANCH=none
BUG=none
TEST=no kernel crash when v3 only device loads

Change-Id: I392db6a7657a5be0f74280ac852f14ec0cd3305c
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1555214
Reviewed-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1568880