CHROMIUM: scripts: Use savedefconfig to minimize kernel config
Rather than saving and then splitting the entire kernel configs, use an
extra "make savedefconfig" step to minimize the result against the
defaults.

Note: This required a slight modification from the 4.19 version. The 4.19
version passed in "$@" to get the LD=, CC=, CXX= pieces. The older version
doesn't have this, and ends up trying to build a target of x86_64 if "$@"
is left in there.

BUG=chromium:963478
TEST=Use CL:1625874 to verify that the total configs haven't changed.

Change-Id: I9ff323f9b7f047a4825ee5a6cdd6c8fe5e33d1a1
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1625877
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>