CHROMIUM: Bluetooth: Enable Bluetooth LE splitter for x86-64
With this change, BT splitter can be used and newblue can be run. By
default the functionality of the splitter is disabled (CL:1352383) and
users can enable it by using the crosh command (CL:1121542).

For now x86-64 is prioritized and this will be enabled on all other
architectures as well later on.

Normalized with: chromeos/scripts/kernelconfig oldconfig

BUG=chromium:859239,chromium:909045
TEST=With this built on caroline, enable and disable newblue and check
that splitter is indeed enabled and disabled. Check that by default
(splitter disabled) Bluetooth works normally.

Change-Id: Ibbe937a113b12e9c3a15ecec6a6c3ac017b6d682
Signed-off-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1575098
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>