FROMLIST: BACKPORT: platform/chrome: wilco_ec: Add telemetry char device interface
The Wilco Embedded Controller is able to send telemetry data
which is useful for enterprise applications. A daemon running on
the OS sends a command to the EC via a write() to a char device,
and can read the response with a read(). The write() request is
verified by the driver to ensure that it is performing only one
of the whitelisted commands, and that no extraneous data is
being transmitted to the EC. The response is passed directly
back to the reader with no modification.

The character device will appear as /dev/wilco_telemN, where N
is some small non-negative integer, starting with 0. Only one
process may have the file descriptor open at a time. The calling
userspace program needs to keep the device file descriptor open
between the calls to write() and read() in order to preserve the
response. 32 bytes of data are expected for arguments, and 32
bytes will be available for reading.

For testing purposes, try requesting the EC's firmware build
date, by sending the WILCO_EC_TELEM_GET_VERSION command with
argument index=3. i.e. write [0x38, 0x00, 0x03, ...(29 more 0s)]
to the device node. An ASCII string of the build date is
returned.

Signed-off-by: Nick Crews <ncrews@chromium.org>

(am from https://lore.kernel.org/patchwork/patch/1070850/)

BUG=b:119046283, b:131692912
TEST=See above

Change-Id: I30000950f511be4a0dc30657e8504376b0e8723e
Reviewed-on: https://chromium-review.googlesource.com/1371034
Commit-Ready: Nick Crews <ncrews@chromium.org>
Tested-by: Nick Crews <ncrews@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>
(cherry picked from commit d9d4c7230369d37466ede72888b9078945cbf882)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1599329
Commit-Queue: Nick Crews <ncrews@chromium.org>