CHROMIUM: drm/i915: Fix PSR exit consistency
PSR on intel seems to have this problem where it doesn't always exit properly.
Instead of displaying the new frame on PSR exit, when we end up doing vblank
evasion we will display the existing frame. The problem is that with double
buffering, the existing frame is also the frame being drawn to, and so we see
partial rendering.

We "fix" this by waiting an extra vblank on modeset when we are transitioning
from PSR enabled to disabled. This add a frame of latency to exit, but the
alternative is to disable PSR completely.

BUG=chromium:951825
TEST=by hand on eve, run a terminal, press enter, don't see any blinking

Change-Id: I2b3d8c272d05a99c646608a9df45c092515ece67
Reviewed-on: https://chromium-review.googlesource.com/1600742
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>