CHROMIUM: mm: totmaps: split PSS into components
This gives us separate numbers for PSS anon, file, and shmem.
We need this to measure types of usage by process group.

This (or ports of this) applies to all chromium kernels up to
and including 4.4.  For 4.14 and later, we need an equivalent
change to smaps_rollup.  (totmaps is specific to chromium/aosp.)
We are trying to upstream that change.

BUG=chromium:962973
TEST=ran it

Change-Id: If14f3ebf5724dfcd27090eaf3c1bceeb3e17a87a
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1626806
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>