CHROMIUM: mm: totmaps: split PSS into components
This gives us separate numbers for PSS anon, file, and shmem.
We need this to measure types of usage by process group.

This (or ports of this) applies to all chromium kernels up to
and including 4.4.  For 4.14 and later, we need an equivalent
change to smaps_rollup.  (totmaps is specific to chromium/aosp.)
We are trying to upstream that change.

BUG=chromium:962973
TEST=ran it

Change-Id: Ia4a677cd64a3f50b231cf298a268b6efe55b41fe
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1620554
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yu Zhao <yuzhao@chromium.org>