CHROMIUM: iwl7000: enable twt by default
In a previous patch, a flag for disabling twt by a dbg-cfg
was added. This will be useful for disabling twt on specific
cores, after branch-out.
However, we'd like twt to be enabled on master.

So, enable twt on master.

Change-Id: I181fc5c80988e7ecea281c60e57a22ae7ec86716
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
iwl7000-tree: d4582e1fdb9d337ff0faa81bc34cd85af82bb3eb