CHROMIUM: iwl7000: pcie: fix ALIVE interrupt handling for gen2 devices w/o MSI-X
We added code to restock the buffer upon ALIVE interrupt
when MSI-X is disabled. This was added as part of the context
info code. This code was added only if the ISR debug level
is set which is very unlikely to be related.
Move this code to run even when the ISR debug level is not
set.

Note that gen2 devices work with MSI-X in most cases so that
this path is seldom used.

Change-Id: Ie37dd13bffce676e94b200f51e0b0ed0495884e0
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: e7242ad3091313de6b5838b666e22f9dc4514e35