CHROMIUM: media: staging/intel-ipu3: sync ipu3-v4l2 to upstream
Currently, in imgu_subdev_open, it use a 1080p crop and fixed format
code, it does not make sense, this patch use the crop and format info
from input node as the initial value, and it also be aligned with
the upstream code.
This patch also remove the initial colorspace and unnecessary
symbol exports.

BUG=b:139647343
TEST=Sanity checked basic camera functions.

Change-Id: Ia41b063b04039987bfb1216818fc10a7a0dff43d
Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1624633
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>