CHROMIUM: blk-mq: Freeze queues independently on hotplug
The code had an optimization to try and freeze all queues in parallel
to minimize the wall time spent on queue reinitialization.
Unfortunately, this can lead to deadlocks when one queue requires
requests on another queue to be processed before it can run dry and
the queues get frozen in the wrong order. This can happen with loop
devices that depend on requests to the underlying device to be
serviced before the loop requests can complete.

BUG=b:132722933
TEST=See bug.

Signed-off-by: Mattias Nissler <mnissler@chromium.org>

Change-Id: Ifa843e63da1987a8f696b50347d5fd3ba58c52f8
Reviewed-on: https://chromium-review.googlesource.com/1626868
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>