CHROMIUM: drm/i915: Disable boost and waitboost
The GT frequency should ideally scale as per the load,
but on making a Hangouts call or opening the camera
preview it is pegged to max frequency. This leads to higher than
expected power consumption(2W+).

The camera stack is probably doing some heavy handed sync operations
which causes the GPU to be always stuck at max frequency.
This is a temporary fix until the userspace(camera) is either
fixed or optimized because we may incur a performance penalty due to
this change.

This bug is discussed here[1].

1. https://bugs.freedesktop.org/show_bug.cgi?id=110750

BUG=b:130638275
TEST=Build and boot on octopus build is successful.

Change-Id: I802a9b0a1f5883e20eabcef0b832a97c495fc014
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1633381
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>