CHROMIUM: iwl7000: mvm: disable TX-AMSDU on older NICs
On older NICs, we occasionally see issues with A-MSDU support,
where the commands in the FIFO get confused and then we see an
assert EDC because the next command in the FIFO isn't TX.

We've tried to isolate this issue and understand where it comes
from, but haven't found any errors in building the A-MSDU in
software.

At least for now, disable A-MSDU support on older hardware so
that users can use it again without fearing the assert.

This fixes https://bugzilla.kernel.org/show_bug.cgi?id=203315.

Change-Id: I2b6a526d09048c8885c472e5797bd4ef25dba360
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: ee4df1794a421f1682d1248f068590887d267da8