CHROMIUM: drm/i915: Disable boost and waitboost
The GT frequency should ideally scale as per the load,
but on making a Hangouts call or opening the camera
preview it is pegged to max frequency. This leads to higher than
expected power consumption(2W+).

The camera stack is probably doing some heavy handed sync operations
which causes the GPU to be always stuck at max frequency.
This is a temporary fix until the userspace(camera) is either
fixed or optimized because we may incur a performance penalty due to
this change.

This bug is discussed here[1].

1. https://bugs.freedesktop.org/show_bug.cgi?id=110750

BUG=b:130638275, b:131723850
TEST=Build and boot on nocturne. Make sure GPU frequency is low when
     running camera app.

Change-Id: I662fada3093ccd5a3424b90dca48deb176a9c313
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1657492
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>